fall-detection-ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
scipy==1.11.4
websockets==12.0
asyncio==3.4.3